+++ .github/workflows/clang-format.yaml
@@ -10,5 +10,5 @@ jobs:
       uses: jidicula/clang-format-action@v4.11.0
       with:
         clang-format-version: '13'
-        check-path: 'src'
+        check-path: 'transport'
         fallback-style: 'Google' # optional
