+++ examples/htool_raw_host_command.c
@@ -17,7 +17,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-#include "../libhoth.h"
+#include "../transports/libhoth_device.h"
 #include "ec_util.h"
 #include "host_commands.h"
 #include "htool.h"
