+++ meson.build
@@ -1,27 +1,16 @@
 project('libhoth', 'c', 'cpp', license: 'Apache-2.0', version: '0.0.0')
 
-libhoth_srcs = [
-  'libhoth.c',
-  'libhoth_mtd.c',
-  'libhoth_usb.c',
-  'libhoth_spi.c',
-  'libhoth_usb_fifo.c',
-  'libhoth_usb_mailbox.c',
-]
+subdir('transports')
 
-libsystemd = dependency('libsystemd')
-libhoth_dbus = static_library('hoth_dbus', 'libhoth_dbus.c', dependencies: [libsystemd])
+install_headers('transports/libhoth_device.h')
+install_headers('transports/libhoth_usb.h')
+install_headers('transports/libhoth_usb_device.h')
+install_headers('libhoth.h')
 
-libusb = dependency('libusb-1.0')
-
-libhoth = library('hoth', libhoth_srcs, dependencies: [libusb], install: true, \
+libhoth = library('hoth', ['libhoth.c'], link_with: [libhoth_transport], dependencies: [libusb], \
     version: meson.project_version())
 
-install_headers('libhoth.h')
-install_headers('libhoth_usb.h')
-install_headers('libhoth_usb_device.h')
-
 pkg = import('pkgconfig')
-pkg.generate(libhoth, name: 'libhoth', description: 'Hoth interface library')
+pkg.generate([libhoth], name: 'libhoth', description: 'Hoth interface library')
 
 subdir('examples')
