+++ transports/libhoth_usb_mailbox.c
@@ -19,7 +19,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "libhoth.h"
+#include "libhoth_device.h"
 #include "libhoth_ec.h"
 #include "libhoth_usb_device.h"
